requirements.txt
biopython>=1.79
numpy>=1.20.0
click>=8.0.0
pytest>=6.0.0
matplotlib>=3.5.0
pybloom-live>=3.0.0